LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/calc-in-font-variation-settings-expected.txt


FAIL calc() specified value serialization in font-variation-settings assert_equals: expected "\"XHGT\" calc(0.7)" but got ""
PASS calc() computed value serialization in font-variation-settings
